Take into account already executed, but skipped modules when determining the "last" module
This is important when executing the "analyze" goal in particular: if we
ignore skipped modules, and the project skips the jqassistant plugin in
the very first module for example,  we will never consider that we are
executing the last module, and thus we will never execute the "analyze"
goal.

We manage to take skipped modules into account by using a simpler
approach to determine whether the current module is the last one.