Do not use a custom class for keys in a map shared across multiple classloaders
Using a custom class for such keys means that executions relying on
different classloaders will use different class definitions and will
create keys that are deemed different even if their content is the same
(because their class is different at runtime).

Since the goal can reasonably be expected not to contain any "@"
character, using simple strings as keys instead of a custom ExecutionKey
class will do.

I am not sure when the various executions of the plugin in different
modules may use different classloaders, but I know this can happen, I've
seen it.

Before this patch, a workaround was to use the
<extensions>true</extensions> option in the jqassistant plugin's
configuration, so that jqassistant classes are added to the core Maven
classloader. But that only gets you so far, as using this workaround
seems to break some features, such as calling goals directly from the
commandline ("mvn jqassistant:server" for example).